+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Option `buildAll` for Suggesters
-- NoResponseRequest Plugin
+- NoWaitForResponseRequest Plugin
 
 ### Fixed
 - PHP 8.2 deprecations for Solarium\QueryType\Server\Collections results